requirements.txt
crewai==1.6.1
crewai-tools==1.6.1
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.13
python-dotenv>=1.1.1
streamlit==1.40.2
